pyproject.toml
[project]
name = "ai_playground"
version = "0.0.2"
readme = "README.md"
dependencies = [
    "docarray>=0.41.0",
    "google-adk>=1.4.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.24",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "pygame>=2.6.1",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "pyyaml>=6.0.2",
    "ruff>=0.5.5",
]
requires-python = ">=3.12"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "src"
